# backend_express/swagger.yaml
openapi: 3.0.3
info:
  title: VPP Aggregation Platform API
  description: |
    API documentation for the Virtual Power Plant (VPP) Aggregation Platform.
    This platform manages edge nodes, telemetry data, dispatch commands, market trading, and multi-agent coordination.
  version: 1.0.0
  contact:
    name: API Support
    email: amansharma12607@gmail.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.vpp-platform.com/api
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/vendor-auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    MarketBid:
      type: object
      required:
        - service_type
        - capacity_mw
        - price_per_mw
      properties:
        service_type:
          type: string
        capacity_mw:
          type: number
        price_per_mw:
          type: number
        duration_minutes:
          type: integer
        market_type:
          type: string
          enum: [intraday, day_ahead, real_time]
    Bid:
      type: object
      properties:
        bid_id:
          type: string
        service_type:
          type: string
        capacity_mw:
          type: number
        price_per_mw:
          type: number
        duration_minutes:
          type: integer
        market_type:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected, cancelled, completed]
        vendor_id:
          type: string
        vendor_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        transaction_id:
          type: string
        bid_id:
          type: string
        vendor_id:
          type: string
        capacity_mw:
          type: number
        price_per_mw:
          type: number
        duration_minutes:
          type: integer
        market_type:
          type: string
        total_amount:
          type: number
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        placed_by:
          type: string
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...
paths:
...
  /market/bids:
    post:
      tags:
        - Market
      summary: Prepare and submit market bid
      description: Create a new market bid
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketBid'
      responses:
        '200':
          description: Bid created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Market
      summary: Get active bids
      description: Retrieve all active market bids
      security: []
      responses:
        '200':
          description: Active bids retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
  /market/bids/{bidId}:
    get:
      tags:
        - Market
      summary: Get bid status
      description: Get status of a specific bid
      security: []
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: string
          description: Bid ID
      responses:
        '200':
          description: Bid status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '404':
          description: Bid not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /marketplace/bids:
    get:
      tags:
        - Marketplace
      summary: Get all bids
      description: Get all marketplace bids with optional filtering
      security: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: market_type
          in: query
          schema:
            type: string
          description: Filter by market type
        - name: placed_by
          in: query
          schema:
            type: string
          description: Filter by user who placed bid
        - name: vendor_id
          in: query
          schema:
            type: string
          description: Filter by vendor ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of records
      responses:
        '200':
          description: Bids retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bid'

    post:
      tags:
        - Marketplace
      summary: Create new bid
      description: Create a new marketplace bid (vendor auth required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarketBid'
      responses:
        '201':
          description: Bid created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Bid'
        '403':
          description: Only vendors can create bids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /marketplace/bids/{bidId}:
    put:
      tags:
        - Marketplace
      summary: Update bid
      description: Update a marketplace bid (vendor auth required, own bids only)
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: string
          description: Bid ID
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                capacity_mw:
                  type: number
                price_per_mw:
                  type: number
                duration_minutes:
                  type: integer
                status:
                  type: string
                  enum: [pending, accepted, rejected, cancelled]
      responses:
        '200':
          description: Bid updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Bid'
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Marketplace
      summary: Delete bid
      description: Delete a marketplace bid (vendor auth required, own bids only)
      parameters:
        - name: bidId
          in: path
          required: true
          schema:
            type: string
          description: Bid ID
      responses:
        '200':
          description: Bid deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '403':
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /marketplace/transactions:
    get:
      tags:
        - Marketplace
      summary: Get all transactions
      description: Get all marketplace transactions with optional filtering
      security: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: vendor_id
          in: query
          schema:
            type: string
          description: Filter by vendor ID
        - name: placed_by
          in: query
          schema:
            type: string
          description: Filter by user who placed transaction
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of records
      responses:
        '200':
          description: Transactions retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'

    post:
      tags:
        - Marketplace
      summary: Create transaction
      description: Create a new transaction (when bid is accepted)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bid_id
                - capacity_mw
                - price_per_mw
              properties:
                bid_id:
                  type: string
                vendor_id:
                  type: string
                capacity_mw:
                  type: number
                price_per_mw:
                  type: number
                duration_minutes:
                  type: integer
                market_type:
                  type: string
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Transaction'
  /marketplace/transactions/{transactionId}:
    put:
      tags:
        - Marketplace
      summary: Update transaction status
      description: Update the status of a transaction
      security: []
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
          description: Transaction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, active, completed, cancelled]
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Transaction'
...
  /marketplace/trading-history:
    get:
      tags:
        - Marketplace
      summary: Get trading history
      description: Get trading history with optional filtering
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
          description: Maximum number of records
      responses:
        '200':
          description: Trading history retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
...